Add analytics calls
