+++ src/app/providers.tsx
@@ -4,8 +4,6 @@ import { WagmiConfig } from 'wagmi'
 
 import { CacheProvider } from '@chakra-ui/next-js'
 import { ChakraProvider, ColorModeScript } from '@chakra-ui/react'
-import { GTMProvider } from '@elgorditosalsero/react-gtm-hook'
-import { ArcxAnalyticsProvider } from '@arcxmoney/analytics'
 import { RainbowKitProvider } from '@rainbow-me/rainbowkit'
 
 import theme, { rainbowkitTheme } from '@/lib/styles/theme'
@@ -13,18 +11,15 @@ import { chains, wagmiConfig } from '@/lib/utils/wagmi'
 
 import '@rainbow-me/rainbowkit/styles.css'
 import '@/lib/styles/fonts'
+import { AnalyticsProvider } from './analytics-provider'
 
 const rainbowKitAppInfo = {
   appName: 'Index Coop',
   learnMoreUrl: 'https://indexcoop.com',
 }
 
-const arcxApiKey = process.env.NEXT_PUBLIC_ARCX_ANALYTICS_API_KEY ?? ''
 
 export function Providers({ children }: { children: React.ReactNode }) {
-  const gtmParams = {
-    id: process.env.NEXT_PUBLIC_GOOGLE_TAG_MANAGER_CONTAINER_ID ?? '',
-  }
   return (
     <CacheProvider prepend={true}>
       <ChakraProvider theme={theme}>
@@ -35,9 +30,9 @@ export function Providers({ children }: { children: React.ReactNode }) {
             theme={rainbowkitTheme}
             appInfo={rainbowKitAppInfo}
           >
-            <ArcxAnalyticsProvider apiKey={arcxApiKey}>
-              <GTMProvider state={gtmParams}>{children}</GTMProvider>
-            </ArcxAnalyticsProvider>
+            <AnalyticsProvider>
+              {children}
+            </AnalyticsProvider>
           </RainbowKitProvider>
         </WagmiConfig>
       </ChakraProvider>
