+++ src/components/footer/footer-link.tsx
@@ -1,4 +1,7 @@
+'use client'
+
 import Link from 'next/link'
+import { useAnalytics } from '@/lib/hooks/use-analytics'
 
 type Props = {
   children: string
@@ -6,12 +9,16 @@ type Props = {
 }
 
 export function FooterLink({ children, href }: Props) {
+  const { logEvent } = useAnalytics()
   return (
     <li>
       <Link
         target='_blank'
         href={href}
         className='text-ic-gray-600 hover:text-ic-gray-900 dark:hover:text-ic-gray-400 dark:text-ic-gray-200 text-sm leading-6'
+        onClick={() =>
+          logEvent('Header Link Clicked', { href })
+        }
       >
         {children}
       </Link>
