+++ src/lib/hooks/use-analytics.ts
@@ -2,7 +2,7 @@ import { useArcxAnalytics } from '@arcxmoney/analytics'
 import { useGTMDispatch } from '@elgorditosalsero/react-gtm-hook'
 import { useCallback } from 'react'
 
-const isProduction = process.env.NEXT_PUBLIC_VERCEL_ENV === 'production'
+const isProduction = process.env.NEXT_PUBLIC_VERCEL_ENV === 'index-app-prod'
 
 export const useAnalytics = () => {
   const arcxSdk = useArcxAnalytics()
@@ -14,6 +14,11 @@ export const useAnalytics = () => {
 
       arcxSdk?.event(name, data)
       sendDataToGTM({ event: name, ...data })
+      window.safary?.track({
+        eventType: 'Generic',
+        eventName: name,
+        parameters: data
+      })
     },
     [arcxSdk, sendDataToGTM]
   )
